+++ examples/training.rb
@@ -4,13 +4,13 @@ require "secryst-trainer"
 
 kh2en = Secryst::Trainer.new(
   model: 'transformer',
-  data_input: File.expand_path('../data/kh-rom/input.csv', __dir__),
-  data_target: File.expand_path('../data/kh-rom/target.csv', __dir__),
+  data_input: File.expand_path('../data/kh-rom-small/input.csv', __dir__),
+  data_target: File.expand_path('../data/kh-rom-small/target.csv', __dir__),
   batch_size: 32,
   lr: 0.1,
   scheduler_step_size: 70,
   gamma: 0.9,
-  max_epochs: nil,
+  max_epochs: 500, # nil for unlimited
   checkpoint_every: 50,
   checkpoint_dir: File.expand_path('./checkpoints', __dir__),
   hyperparameters: {
