+++ examples/translating.rb
@@ -5,7 +5,7 @@ require "secryst"
 kh2en = Secryst::Translator.new({
   model: 'transformer',
   vocabs_dir: File.expand_path('./checkpoints/', __dir__),
-  model_file: File.expand_path('./checkpoints/checkpoint-150.pth', __dir__),
+  model_file: File.expand_path('./checkpoints/checkpoint-500.pth', __dir__),
   hyperparameters: {
     d_model: 64,
     nhead: 8,
