+++ lib/secryst-trainer.rb
@@ -7,9 +7,9 @@ require "numo/linalg/use/openblas"
 require "torch-rb"
 
 # transformer model
-require "secryst/models/transformer/clip_grad_norm"
-require "secryst/models/transformer/multihead_attention"
-require "secryst/models/transformer/vocab"
-require "secryst/models/transformer"
+require "secryst/clip_grad_norm"
+require "secryst/multihead_attention"
+require "secryst/vocab"
+require "secryst/transformer"
 
 require "secryst/trainer"
