+++ lib/secryst.rb
@@ -4,8 +4,8 @@ require 'json'
 require "torch-rb"
 
 # transformer model
-require "secryst/models/transformer/multihead_attention"
-require "secryst/models/transformer/vocab"
-require "secryst/models/transformer"
+require "secryst/multihead_attention"
+require "secryst/vocab"
+require "secryst/transformer"
 
 require "secryst/translator"
