+++ lib/secryst/vocab.rb
